add abstract scanner class for read repair use cases
